Fix multi-parameter provider issue in core, spark and flink
