+++ core/src/test/java/org/apache/iceberg/TestCreateSnapshotEvent.java
@@ -20,6 +20,8 @@ package org.apache.iceberg;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Arrays;
+import java.util.List;
 import org.apache.iceberg.events.CreateSnapshotEvent;
 import org.apache.iceberg.events.Listener;
 import org.apache.iceberg.events.Listeners;
@@ -29,6 +31,11 @@ import org.junit.jupiter.api.TestTemplate;
 public class TestCreateSnapshotEvent extends TestBase {
 
   private CreateSnapshotEvent currentEvent;
+
+  @Parameters(name = "formatVersion = {0}")
+  protected static List<Object> parameters() {
+    return Arrays.asList(1, 2, 3);
+  }
 
   @BeforeEach
   public void initListener() {
