+++ flink/v1.18/build.gradle
@@ -123,6 +123,11 @@ project(":iceberg-flink:iceberg-flink-${flinkMajorVersion}") {
 
   test {
     useJUnitPlatform()
+
+    // This flag is set to keep the backward compatibility between JUnit 5.10 and 5.11.
+    // Remove this after all Spark tests are migrated to JUnit 5.11.
+    // See https://github.com/apache/iceberg/pull/11262#issuecomment-2395411961
+    jvmArgs += "-Djunit.platform.reflection.search.useLegacySemantics=true"
   }
 }
 
@@ -255,6 +260,11 @@ project(":iceberg-flink:iceberg-flink-runtime-${flinkMajorVersion}") {
     description = "Test Flink Runtime Jar against Flink ${flinkMajorVersion}"
     group = "verification"
     jvmArgs += project.property('extraJvmArgs')
+
+    // This flag is set to keep the backward compatibility between JUnit 5.10 and 5.11.
+    // Remove this after all Spark tests are migrated to JUnit 5.11.
+    // See https://github.com/apache/iceberg/pull/11262#issuecomment-2395411961
+    jvmArgs += "-Djunit.platform.reflection.search.useLegacySemantics=true"
     testClassesDirs = sourceSets.integration.output.classesDirs
     classpath = sourceSets.integration.runtimeClasspath + files(shadowJar.archiveFile.get().asFile.path)
     inputs.file(shadowJar.archiveFile.get().asFile.path)
