+++ spark/v3.5/spark/src/test/java/org/apache/iceberg/spark/sql/TestUnpartitionedWritesToBranch.java
@@ -22,6 +22,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.apache.iceberg.Table;
 import org.apache.iceberg.exceptions.ValidationException;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestTemplate;
 
 public class TestUnpartitionedWritesToBranch extends UnpartitionedWritesTestBase {
@@ -29,6 +30,7 @@ public class TestUnpartitionedWritesToBranch extends UnpartitionedWritesTestBase
   private static final String BRANCH = "test";
 
   @Override
+  @BeforeEach
   public void createTables() {
     super.createTables();
     Table table = validationCatalog.loadTable(tableIdent);
